put the correct file in the correct place
